config rejigs + added a bunch of new options fixed fluid physics disparity between loaders (fucking finally omgggg) fluid now teleports you instantly when you become fully submerged in it (toggleable) removed enchantment glint from dimtears bottle (looks better imo--commented out so we can easily add it back later) renamed dimtears bottle (no mo' bottle o') rift flame particles now properly adhere to the surface of the fluid improved a couple of mixins rift flame particle no longer inherits an original class. there's literally a freaking method for defining a billboard type

+++ common/src/main/java/com/ordana/dimensional_tears/mixins/nether_portal/NetherPortalBlockMixin.java
@@ -31,9 +31,9 @@ public class NetherPortalBlockMixin {
         Direction.Axis axis = blockState.getValue(AXIS);
         PortalShape portalShape = new PortalShape(levelAccessor, blockPos, axis);
 
-        float chance = CommonConfigs.PORTAL_DESTRUCTION_CRYING_OBSIDIAN_CHANCE.get();
+        double chance = CommonConfigs.PORTAL_DESTRUCTION_CRYING_OBSIDIAN_CHANCE.get();
 
-        if (portalShape.isComplete() || chance <= 0)
+        if (portalShape.isComplete() || chance <= 0.0)
             return;
 
         RandomSource random = levelAccessor.getRandom();
